+++ home/scripts/void.sh
@@ -37,7 +37,7 @@ add_alsa() {
 }
 
 add_pipewire_alsa() {
-    pkgs="$pkgs pipewire alsa-pipewire wireplumber wireplumber-elogind wiremix"
+    pkgs="$pkgs pipewire alsa-pipewire wireplumber wireplumber-elogind wiremix yazi Thunar tumbler"
 }
 
 add_swayfx() {
@@ -45,11 +45,11 @@ add_swayfx() {
     inotify-tools kitty ImageMagick cargo
     playerctl mpv-mpris eww swaybg grim
     iwd elogind mesa-dri dunst chafa wlr-randr
-    swayidle swaylock dbus brightnessctl"
+    swayidle swaylock dbus brightnessctl kew
+    glances element-desktop fd"
 
     #user_groups="$user_groups _seatd"
-
-    services="$services dbus"
+    services="$services dbus elogind"
 }
 
 # No brave package...